Update test coverage for Caregivers app helper utils
