+++ src/applications/caregivers/definitions/sharedUI.js
@@ -1,5 +1,4 @@
 import { merge } from 'lodash';
-import get from 'platform/utilities/data/get';
 import VaTextInputField from 'platform/forms-system/src/js/web-component-fields/VaTextInputField';
 import {
   radioUI,
@@ -13,6 +12,7 @@ import {
 } from 'platform/forms-system/src/js/web-component-patterns';
 import { genderLabels } from 'platform/static-data/labels';
 import { validateSsnIsUnique, requireAddressFields } from '../utils/validation';
+import { setAddressCountry } from '../utils/helpers/schema';
 import { replaceStrValues } from '../utils/helpers';
 import AddressWithAutofill from '../components/FormFields/AddressWithAutofill';
 import CustomReviewField from '../components/FormReview/CustomReviewField';
@@ -29,19 +29,7 @@ export const addressUI = props => {
     country: {
       'ui:options': {
         hideOnReview: true,
-        updateSchema: (formData, _schema, uiSchema, _index, path) => {
-          const USA = { value: 'USA', label: 'United States' };
-          const addressPath = path.slice(0, -1);
-          const countryUI = uiSchema;
-          const addressFormData = get(addressPath, formData) ?? {};
-          countryUI['ui:options'].inert = true;
-          addressFormData.country = USA.value;
-          return {
-            enum: [USA.value],
-            enumNames: [USA.label],
-            default: USA.value,
-          };
-        },
+        updateSchema: setAddressCountry,
       },
     },
     street: {
