+++ src/applications/caregivers/tests/unit/config/secondaryTwo/mailingAddress.unit.spec.js
@@ -1,0 +1,34 @@
+import {
+  testNumberOfErrorsOnSubmitForWebComponents,
+  testNumberOfWebComponentFields,
+} from '../helpers.spec';
+import formConfig from '../../../../config/form';
+
+const {
+  chapters: {
+    secondaryCaregiverInformation: {
+      pages: { secondaryTwoMailingAddress },
+    },
+  },
+} = formConfig;
+const { title: pageTitle, schema, uiSchema } = secondaryTwoMailingAddress;
+
+// run test for correct number of fields on the page
+const expectedNumberOfWebComponentFields = 7;
+testNumberOfWebComponentFields(
+  formConfig,
+  schema,
+  uiSchema,
+  expectedNumberOfWebComponentFields,
+  pageTitle,
+);
+
+// run test for correct number of error messages on submit
+const expectedNumberOfWebComponentErrors = 3;
+testNumberOfErrorsOnSubmitForWebComponents(
+  formConfig,
+  schema,
+  uiSchema,
+  expectedNumberOfWebComponentErrors,
+  pageTitle,
+);
