+++ src/applications/caregivers/tests/unit/config/veteran/homeAddress.unit.spec.js
@@ -1,0 +1,34 @@
+import {
+  testNumberOfErrorsOnSubmitForWebComponents,
+  testNumberOfWebComponentFields,
+} from '../helpers.spec';
+import formConfig from '../../../../config/form';
+
+const {
+  chapters: {
+    veteranInformation: {
+      pages: { vetHomeAddress },
+    },
+  },
+} = formConfig;
+const { title: pageTitle, schema, uiSchema } = vetHomeAddress;
+
+// run test for correct number of fields on the page
+const expectedNumberOfWebComponentFields = 7;
+testNumberOfWebComponentFields(
+  formConfig,
+  schema,
+  uiSchema,
+  expectedNumberOfWebComponentFields,
+  pageTitle,
+);
+
+// run test for correct number of error messages on submit
+const expectedNumberOfWebComponentErrors = 4;
+testNumberOfErrorsOnSubmitForWebComponents(
+  formConfig,
+  schema,
+  uiSchema,
+  expectedNumberOfWebComponentErrors,
+  pageTitle,
+);
